requirements.txt
Flask==3.0.3 
pdfminer.six==20231228
python-docx==1.1.2
sentence-transformers==3.0.1
scikit-learn==1.4.1.post1
pandas==2.2.1 
spacy==3.7.5
pydantic==2.8.2